Added gbsearch function
It's kind of a pain to search each local git branch manually, so why not
make a bash function to do it all for me?  gbsearch is just that, added
simple function impl along with necessary pathing changes to bashrc.  At
some point, these all need to be converted to python utils.